Test Keryx.Core exhaustively; harden ByteWriter.Patch bounds check
123 tests covering every reader/writer width against hand-computed
big-endian vectors, truncation at every boundary, Reserve/Patch
back-patching and logger semantics. The tests surfaced a theoretical
integer-overflow bypass in Patch's offset+count bound; compute it in
64-bit so adversarial windows fail with ByteBufferException as documented.

+++ src/Keryx.Core/ByteWriter.cs
@@ -115,7 +115,7 @@ public ref struct ByteWriter
     /// <summary>Returns the <paramref name="count"/>-byte window at <paramref name="offset"/> for back-patching.</summary>
     public readonly Span<byte> Patch(int offset, int count)
     {
-        if (offset < 0 || count < 0 || offset + count > _position)
+        if (offset < 0 || count < 0 || (long)offset + count > _position)
         {
             throw new ByteBufferException(
                 $"Patch window [{offset}, {offset + count}) is outside the written range [0, {_position}).");
